cache visit elements
